refactor: change query bson to fluent api

+++ src/ChatService.Contract/Abstractions/Shared/PagedList.cs
@@ -3,12 +3,12 @@
 public class PagedList<T> 
 {
     public List<T> Items { get; set; }
-    public int TotalItems { get; init; }
+    public long TotalItems { get; init; }
     public int PageSize { get; init; }
     public string NextCursor { get; init; }
     public bool HasNextPage { get; init; }
 
-    private PagedList(List<T> items, int total, int pageSize, string nextCursor, bool hasNextPage)
+    private PagedList(List<T> items, long total, int pageSize, string nextCursor, bool hasNextPage)
     {
         Items = items;
         TotalItems = total;
@@ -17,7 +17,7 @@ public class PagedList<T>
         HasNextPage = hasNextPage;
     }
     
-    public static PagedList<T> Create(List<T> items, int total, int pageSize, string nextCursor, bool hasNext)
+    public static PagedList<T> Create(List<T> items, long total, int pageSize, string nextCursor, bool hasNext)
         => new(items, total, pageSize, nextCursor, hasNext);
     
 }
